#719 refactored sitemapcontroller to API, fixed unittests that called SwiftProvier bean
