--- api2-model/src/main/java/eu/europeana/api2/model/enums/Profile.java
@@ -1,15 +1,0 @@
-package eu.europeana.api2.model.enums;
-
-public enum Profile {
-	MINIMAL("minimal"), STANDARD("standard"), RICH("rich"), SIMILAR("similar");
-
-	private String name;
-
-	Profile(String name) {
-		this.name = name;
-	}
-
-	public String getName() {
-		return name;
-	}
-}
