+++ api2-model/src/main/java/eu/europeana/api2/v2/model/enums/DefaultImage.java
